+++ sql/src/main/java/org/apache/druid/sql/calcite/planner/querygen/Vertex.java
@@ -20,7 +20,8 @@
 package org.apache.druid.sql.calcite.planner.querygen;
 
 import org.apache.calcite.rel.RelNode;
-import org.apache.druid.sql.calcite.planner.DruidQueryGenerator.InputDesc;
+import org.apache.druid.query.DataSource;
+import org.apache.druid.segment.column.RowSignature;
 import org.apache.druid.sql.calcite.rel.DruidQuery;
 import org.apache.druid.sql.calcite.rel.PartialDruidQuery;
 
@@ -31,6 +32,21 @@ import org.apache.druid.sql.calcite.rel.PartialDruidQuery;
  */
 public interface Vertex
 {
+
+  /**
+   * Utility class to return represent input related things.
+   */
+  public class InputDesc
+  {
+    public DataSource dataSource;
+    public RowSignature rowSignature;
+
+    public InputDesc(DataSource dataSource, RowSignature rowSignature)
+    {
+      this.dataSource = dataSource;
+      this.rowSignature = rowSignature;
+    }
+  }
   boolean canUnwrapInput();
 
   InputDesc unwrapInputDesc();
